Move back ShapeTest into masking_test.py to simplify diff
